Moved inventory2 and gcp_api2 packages under iam/inventory
