+++ google/cloud/security/iam/inventory/gcp_api2/compute.py
@@ -17,7 +17,7 @@
 import os
 from ratelimiter import RateLimiter
 
-from google.cloud.security.common.gcp_api2 import _base_client
+from google.cloud.security.iam.inventory.gcp_api2 import _base_client
 from google.cloud.security.common.util import log_util
