+++ google/cloud/security/iam/inventory/inventory2/resources.py
