+++ google/cloud/security/iam/inventory/inventory2/storage.py
