+++ tests/iam/gcp_tests/gcp_recorded_test.py
@@ -19,8 +19,8 @@ import unittest
 import os
 
 from tests.unittest_utils import ForsetiTestCase
-from google.cloud.security.inventory2.storage import Memory as MemoryStorage
-from google.cloud.security.inventory2.progress import Progresser
+from google.cloud.security.iam.inventory.inventory2.storage import Memory as MemoryStorage
+from google.cloud.security.iam.inventory.inventory2.progress import Progresser
 from google.cloud.security.iam.inventory.crawler import run_crawler
 from google.cloud.security.iam.server import InventoryConfig
