+++ tests/iam/inventory_tests/crawling_test.py
@@ -18,8 +18,8 @@ import unittest
 
 
 from tests.unittest_utils import ForsetiTestCase
-from google.cloud.security.inventory2.storage import Memory as MemoryStorage
-from google.cloud.security.inventory2.progress import Progresser
+from google.cloud.security.iam.inventory.inventory2.storage import Memory as MemoryStorage
+from google.cloud.security.iam.inventory.inventory2.progress import Progresser
 from google.cloud.security.iam.inventory.crawler import run_crawler
 from google.cloud.security.iam.server import InventoryConfig
 from tests.iam.utils.gcp_env import gcp_configured, gcp_env
